# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NomadFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Trip {
  id: number;
  name: string;
  destination: string;
  startDate: string;
  endDate: string;
  coverColor: string;
  createdAt: string;
}

export type ActivityCategory =
  (typeof ActivityCategory)[keyof typeof ActivityCategory];

export const ActivityCategory = {
  transport: "transport",
  accommodation: "accommodation",
  food: "food",
  nature: "nature",
  culture: "culture",
  shopping: "shopping",
  nightlife: "nightlife",
  other: "other",
} as const;

export interface Activity {
  id: number;
  tripId: number;
  dayIndex: number;
  title: string;
  category: ActivityCategory;
  startTime: string;
  endTime?: string | null;
  location?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type TripDetail = Trip & {
  activities: Activity[];
};

export interface CreateTripInput {
  name: string;
  destination: string;
  startDate: string;
  endDate: string;
  coverColor: string;
}

export type CreateActivityInputCategory =
  (typeof CreateActivityInputCategory)[keyof typeof CreateActivityInputCategory];

export const CreateActivityInputCategory = {
  transport: "transport",
  accommodation: "accommodation",
  food: "food",
  nature: "nature",
  culture: "culture",
  shopping: "shopping",
  nightlife: "nightlife",
  other: "other",
} as const;

export interface CreateActivityInput {
  dayIndex: number;
  title: string;
  category: CreateActivityInputCategory;
  startTime: string;
  endTime?: string | null;
  location?: string | null;
  notes?: string | null;
}

export type UpdateActivityInputCategory =
  (typeof UpdateActivityInputCategory)[keyof typeof UpdateActivityInputCategory];

export const UpdateActivityInputCategory = {
  transport: "transport",
  accommodation: "accommodation",
  food: "food",
  nature: "nature",
  culture: "culture",
  shopping: "shopping",
  nightlife: "nightlife",
  other: "other",
} as const;

export interface UpdateActivityInput {
  title?: string;
  category?: UpdateActivityInputCategory;
  startTime?: string;
  endTime?: string | null;
  location?: string | null;
  notes?: string | null;
}

export type GetActivitiesParams = {
  dayIndex?: number;
};