# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NomadFlow API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: trips
    description: Trip operations
  - name: activities
    description: Activity operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /trips:
    get:
      operationId: getTrips
      tags: [trips]
      summary: List all trips
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripInput"
      responses:
        "201":
          description: Created trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{tripId}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a single trip with its activities
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip with activities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripDetail"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip and all its activities
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Trip deleted
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trips/{tripId}/activities:
    get:
      operationId: getActivities
      tags: [activities]
      summary: List activities for a trip, optionally filtered by day index
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: dayIndex
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create an activity for a trip day
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityInput"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /trips/{tripId}/activities/{activityId}:
    put:
      operationId: updateActivity
      tags: [activities]
      summary: Update an activity
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivityInput"
      responses:
        "200":
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "404":
          description: Activity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete an activity
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: activityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Activity deleted
        "404":
          description: Activity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Trip:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        destination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        coverColor:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - destination
        - startDate
        - endDate
        - coverColor
        - createdAt
    TripDetail:
      allOf:
        - $ref: "#/components/schemas/Trip"
        - type: object
          properties:
            activities:
              type: array
              items:
                $ref: "#/components/schemas/Activity"
          required:
            - activities
    CreateTripInput:
      type: object
      properties:
        name:
          type: string
        destination:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        coverColor:
          type: string
      required:
        - name
        - destination
        - startDate
        - endDate
        - coverColor
    Activity:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        dayIndex:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [transport, accommodation, food, nature, culture, shopping, nightlife, other]
        startTime:
          type: string
        endTime:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tripId
        - dayIndex
        - title
        - category
        - startTime
        - createdAt
    CreateActivityInput:
      type: object
      properties:
        dayIndex:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [transport, accommodation, food, nature, culture, shopping, nightlife, other]
        startTime:
          type: string
        endTime:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - dayIndex
        - title
        - category
        - startTime
    UpdateActivityInput:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
          enum: [transport, accommodation, food, nature, culture, shopping, nightlife, other]
        startTime:
          type: string
        endTime:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true