# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NomadFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityCategory } from "./activityCategory";

export interface Activity {
  id: number;
  tripId: number;
  dayIndex: number;
  title: string;
  category: ActivityCategory;
  startTime: string;
  endTime?: string | null;
  location?: string | null;
  notes?: string | null;
  createdAt: Date;
}