# lib/api-zod/src/generated/types/createActivityInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NomadFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateActivityInputCategory } from "./createActivityInputCategory";

export interface CreateActivityInput {
  dayIndex: number;
  title: string;
  category: CreateActivityInputCategory;
  startTime: string;
  endTime?: string | null;
  location?: string | null;
  notes?: string | null;
}