# lib/api-zod/src/generated/types/createTripInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NomadFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTripInput {
  name: string;
  destination: string;
  startDate: Date;
  endDate: Date;
  coverColor: string;
}