# lib/api-zod/src/generated/types/tripDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NomadFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from "./activity";
import type { Trip } from "./trip";

export type TripDetail = Trip & {
  activities: Activity[];
};