# lib/api-zod/src/generated/types/updateActivityInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NomadFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateActivityInputCategory } from "./updateActivityInputCategory";

export interface UpdateActivityInput {
  title?: string;
  category?: UpdateActivityInputCategory;
  startTime?: string;
  endTime?: string | null;
  location?: string | null;
  notes?: string | null;
}